requirements.txt
singer-python~=5.12.1
google-cloud-bigquery>=2.26.0
jsonschema~=2.6.0
setuptools~=60.3.1
